+++ src/client/gui/client_ui.h
@@ -98,7 +98,6 @@ private Q_SLOTS:
     void menuSlotDeviceDelete() override;
     void menuSlotDeviceImport() override;
     void menuSlotDeviceExport() override;
-    void menuSlotDeviceEnable();
     // ----------------------------
     // ----------DATA VIEW---------
     // ----------------------------
@@ -123,8 +122,6 @@ private Q_SLOTS:
 
 private Q_SLOTS:
     void contextMenuDevice(mbClientDevice *device);
-    void changeCurrentDevice(mbCoreDevice *device);
-    void setDeviceEnabled(bool enable);
 
 private:
     void editPort(mbCorePort *port);