+++ src/server/gui/server_ui.cpp
@@ -141,6 +141,7 @@ mbServerUi::mbServerUi(mbServer *core, QWidget *parent) :
     m_ui.actionDeviceDelete              = ui->actionDeviceDelete             ;
     m_ui.actionDeviceImport              = ui->actionDeviceImport             ;
     m_ui.actionDeviceExport              = ui->actionDeviceExport             ;
+    m_ui.actionDeviceEnable              = ui->actionDeviceEnable             ;
     m_ui.actionDeviceStatistics          = ui->actionDeviceStatistics         ;
     m_ui.actionDataViewItemNew           = ui->actionDataViewItemNew          ;
     m_ui.actionDataViewItemEdit          = ui->actionDataViewItemEdit         ;