+++ changelog.md
@@ -229,7 +229,7 @@ List of changes of new v0.4 version of ModbusTools:
 # 0.5.1
 
 * Added device settings for write function of single coil/register
-* Added enable/disable for mbclient device (menu `Device/Enable`, context menu)
+* Added enable/disable for client/server device (menu `Device/Enable`, context menu)
 * Fixed issue #70: Changing RegisterOrder doesn't affect the value
 * Fixed issue for client's device dialog, when Port type combobox was empty
 * Added `Response Time` statistics for client